requirements.txt
beautifulsoup4==4.8.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
fire==0.4.0
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.8
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.3
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.20.1
panda==0.3.1
pandas==1.2.3
pdfkit==0.6.1
Pillow==8.1.2
ping3==2.6.6
prettytable==2.1.0
pycryptodome==3.10.1
Pygments==2.8.1
pymongo==3.11.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.22.0
six==1.15.0
soupsieve==2.2
termcolor==1.1.0
tqdm==4.59.0
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
wordcloud==1.8.1
XlsxWriter==1.3.7
Zhihu-Spider==6.0.1
zhihuer==0.0.1
